+++ impl/src/fmt/display.rs
@@ -3,16 +3,23 @@
 #[cfg(doc)]
 use std::fmt;
 
+use convert_case::Casing;
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
-use syn::{ext::IdentExt as _, parse_quote, spanned::Spanned as _};
-
-use crate::utils::{attr::ParseMultiple as _, Spanning};
-
-use super::{
-    trait_name_to_attribute_name, ContainerAttributes, ContainsGenericsExt as _,
-    FmtAttribute,
+use syn::{
+    ext::IdentExt as _,
+    parse::{Parse, ParseStream},
+    parse_quote,
+    spanned::Spanned as _,
+    token, LitStr,
 };
+
+use crate::utils::{
+    attr::{self, ParseMultiple as _},
+    Spanning,
+};
+
+use super::{trait_name_to_attribute_name, ContainsGenericsExt as _, FmtAttribute};
 
 /// Expands a [`fmt::Display`]-like derive macro.
 ///
@@ -74,6 +81,178 @@ pub fn expand(input: &syn::DeriveInput, trait_name: &str) -> syn::Result<TokenSt
     })
 }
 
+/// Representation of possible [`fmt::Display`]-like derive macro attributes placed on a container
+/// (struct or enum variant).
+///
+/// ```rust,ignore
+/// #[<attribute>("<fmt-literal>", <fmt-args>)]
+/// #[<attribute>(bound(<where-predicates>))]
+/// #[<attribute>(rename_all = "<casing>")]
+/// ```
+///
+/// `#[<attribute>("...")]` and `#[<attribute>(rename_all = "...")]` can be specified only once,
+/// while multiple `#[<attribute>(bound(...))]` are allowed.
+#[derive(Debug, Default)]
+struct ContainerAttributes {
+    /// [`RenameAllAttribute`] for case convertion.
+    rename_all: Option<RenameAllAttribute>,
+
+    /// Common [`ContainerAttributes`].
+    ///
+    /// [`ContainerAttributes`]: super::ContainerAttributes
+    common: super::ContainerAttributes,
+}
+
+impl Parse for ContainerAttributes {
+    fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
+        mod ident {
+            use syn::custom_keyword;
+
+            custom_keyword!(bounds);
+            custom_keyword!(bound);
+            custom_keyword!(rename_all);
+        }
+
+        // We do check `FmtAttribute::check_legacy_fmt` eagerly here, because `.lookahead1()` won't
+        // check for the `fmt =` ident below, to omit including it into the `ahead.error()`.
+        FmtAttribute::check_legacy_fmt(input)?;
+
+        // We use `.lookahead1()` with all possible idents to form a nice error message including
+        // all the possible variants.
+        let ahead = input.lookahead1();
+        if ahead.peek(LitStr)
+            || ahead.peek(ident::bounds)
+            || ahead.peek(ident::bound)
+            || ahead.peek(token::Where)
+        {
+            Ok(Self {
+                common: input.parse()?,
+                ..Default::default()
+            })
+        } else if ahead.peek(ident::rename_all) {
+            Ok(Self {
+                rename_all: Some(input.parse()?),
+                ..Default::default()
+            })
+        } else {
+            return Err(ahead.error());
+        }
+    }
+}
+
+impl attr::ParseMultiple for ContainerAttributes {
+    fn merge_attrs(
+        prev: Spanning<Self>,
+        new: Spanning<Self>,
+        name: &syn::Ident,
+    ) -> syn::Result<Spanning<Self>> {
+        let Spanning {
+            span: prev_span,
+            item: mut prev,
+        } = prev;
+        let Spanning {
+            span: new_span,
+            item: new,
+        } = new;
+
+        if new
+            .rename_all
+            .and_then(|n| prev.rename_all.replace(n))
+            .is_some()
+        {
+            return Err(syn::Error::new(
+                new_span,
+                format!("multiple `#[{name}(rename_all=\"...\")]` attributes aren't allowed"),
+            ));
+        }
+        prev.common = super::ContainerAttributes::merge_attrs(
+            Spanning::new(prev.common, prev_span),
+            Spanning::new(new.common, new_span),
+            name,
+        )?
+        .into_inner();
+
+        Ok(Spanning::new(
+            prev,
+            prev_span.join(new_span).unwrap_or(prev_span),
+        ))
+    }
+}
+
+/// Representation of a `rename_all` macro attribute.
+///
+/// ```rust,ignore
+/// #[<attribute>(rename_all = "...")]
+/// ```
+///
+/// Possible cases:
+/// - `lowercase`
+/// - `UPPERCASE`
+/// - `PascalCase`
+/// - `camelCase`
+/// - `snake_case`
+/// - `SCREAMING_SNAKE_CASE`
+/// - `kebab-case`
+/// - `SCREAMING-KEBAB-CASE`
+#[derive(Debug, Clone, Copy)]
+enum RenameAllAttribute {
+    Lower,
+    Upper,
+    Pascal,
+    Camel,
+    Snake,
+    ScreamingSnake,
+    Kebab,
+    ScreamingKebab,
+}
+
+impl RenameAllAttribute {
+    /// Converts the provided `name` into the case of this [`RenameAllAttribute`].
+    fn convert_case(&self, name: &str) -> String {
+        let case = match self {
+            Self::Lower => convert_case::Case::Flat,
+            Self::Upper => convert_case::Case::UpperFlat,
+            Self::Pascal => convert_case::Case::Pascal,
+            Self::Camel => convert_case::Case::Camel,
+            Self::Snake => convert_case::Case::Snake,
+            Self::ScreamingSnake => convert_case::Case::UpperSnake,
+            Self::Kebab => convert_case::Case::Kebab,
+            Self::ScreamingKebab => convert_case::Case::UpperKebab,
+        };
+        name.to_case(case)
+    }
+}
+
+impl Parse for RenameAllAttribute {
+    fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
+        let _ = input.parse::<syn::Path>().and_then(|p| {
+            if p.is_ident("rename_all") {
+                Ok(p)
+            } else {
+                Err(syn::Error::new(
+                    p.span(),
+                    "unknown attribute argument, expected `rename_all = \"...\"`",
+                ))
+            }
+        })?;
+
+        input.parse::<token::Eq>()?;
+
+        let value: LitStr = input.parse()?;
+        Ok(match value.value().replace(['-', '_'], "").to_lowercase().as_str() {
+            "lowercase" => Self::Lower,
+            "uppercase" => Self::Upper,
+            "pascalcase" => Self::Pascal,
+            "camelcase" => Self::Camel,
+            "snakecase" => Self::Snake,
+            "screamingsnakecase" => Self::ScreamingSnake,
+            "kebabcase" => Self::Kebab,
+            "screamingkebabcase" => Self::ScreamingKebab,
+            _ => return Err(syn::Error::new_spanned(value, "unexpected casing expected one of: \"lowercase\", \"UPPERCASE\", \"PascalCase\", \"camelCase\", \"snake_case\", \"SCREAMING_SNAKE_CASE\", \"kebab-case\", or \"SCREAMING-KEBAB-CASE\""))
+        })
+    }
+}
+
 /// Type alias for an expansion context:
 /// - [`ContainerAttributes`].
 /// - Type parameters. Slice of [`syn::Ident`].
@@ -130,7 +309,7 @@ fn expand_enum(
     e: &syn::DataEnum,
     (container_attrs, type_params, _, trait_ident, attr_name): ExpansionCtx<'_>,
 ) -> syn::Result<(Vec<syn::WherePredicate>, TokenStream)> {
-    if let Some(shared_fmt) = &container_attrs.fmt {
+    if let Some(shared_fmt) = &container_attrs.common.fmt {
         if shared_fmt
             .placeholders_by_arg("_variant")
             .any(|p| p.has_modifiers || p.trait_name != "Display")
@@ -148,12 +327,12 @@ fn expand_enum(
     let (bounds, match_arms) = e.variants.iter().try_fold(
         (Vec::new(), TokenStream::new()),
         |(mut bounds, mut arms), variant| {
-            let attrs = ContainerAttributes::parse_attrs(&variant.attrs, attr_name)?
+            let mut attrs = ContainerAttributes::parse_attrs(&variant.attrs, attr_name)?
                 .map(Spanning::into_inner)
                 .unwrap_or_default();
             let ident = &variant.ident;
 
-            if attrs.fmt.is_none()
+            if attrs.common.fmt.is_none()
                 && variant.fields.is_empty()
                 && attr_name != "display"
             {
@@ -166,8 +345,12 @@ fn expand_enum(
                 ));
             }
 
+            if let Some(rename_all) = container_attrs.rename_all {
+                attrs.rename_all.get_or_insert(rename_all);
+            }
+
             let v = Expansion {
-                shared_attr: container_attrs.fmt.as_ref(),
+                shared_attr: container_attrs.common.fmt.as_ref(),
                 attrs: &attrs,
                 fields: &variant.fields,
                 type_params,
@@ -210,7 +393,7 @@ fn expand_union(
     u: &syn::DataUnion,
     (attrs, _, _, _, attr_name): ExpansionCtx<'_>,
 ) -> syn::Result<(Vec<syn::WherePredicate>, TokenStream)> {
-    let fmt = &attrs.fmt.as_ref().ok_or_else(|| {
+    let fmt = &attrs.common.fmt.as_ref().ok_or_else(|| {
         syn::Error::new(
             u.fields.span(),
             format!("unions must have `#[{attr_name}(\"...\", ...)]` attribute"),
@@ -218,7 +401,7 @@ fn expand_union(
     })?;
 
     Ok((
-        attrs.bounds.0.clone().into_iter().collect(),
+        attrs.common.bounds.0.clone().into_iter().collect(),
         quote! { derive_more::core::write!(__derive_more_f, #fmt) },
     ))
 }
@@ -285,7 +468,7 @@ impl Expansion<'_> {
 
         let (has_shared_attr, shared_attr_is_wrapping) = self.shared_attr_info();
 
-        let wrap_into_shared_attr = match &self.attrs.fmt {
+        let wrap_into_shared_attr = match &self.attrs.common.fmt {
             Some(fmt) => {
                 body = if shared_attr_is_wrapping {
                     let deref_args = fmt.additional_deref_args(self.fields);
@@ -305,7 +488,10 @@ impl Expansion<'_> {
             None => {
                 if shared_attr_is_wrapping || !has_shared_attr {
                     body = if self.fields.is_empty() {
-                        let ident_str = self.ident.unraw().to_string();
+                        let mut ident_str = self.ident.unraw().to_string();
+                        if let Some(rename_all) = &self.attrs.rename_all {
+                            ident_str = rename_all.convert_case(&ident_str);
+                        }
 
                         if shared_attr_is_wrapping {
                             quote! { #ident_str }
@@ -377,7 +563,7 @@ impl Expansion<'_> {
 
         let (has_shared_attr, shared_attr_is_wrapping) = self.shared_attr_info();
 
-        let mix_shared_attr_bounds = match &self.attrs.fmt {
+        let mix_shared_attr_bounds = match &self.attrs.common.fmt {
             Some(attr) => {
                 bounds.extend(
                     attr.bounded_types(self.fields)
@@ -389,7 +575,7 @@ impl Expansion<'_> {
 
                             Some(parse_quote! { #ty: derive_more::core::fmt::#trait_ident })
                         })
-                        .chain(self.attrs.bounds.0.clone()),
+                        .chain(self.attrs.common.bounds.0.clone()),
                 );
                 shared_attr_is_wrapping
             }