+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Support `#[display(rename_all = "<casing>")]` attribute to change output for
+  implicit naming of unit enum variants or unit structs when deriving `Display`.
+  ([#443](https://github.com/JelteF/derive_more/pull/443))
+
 ## 2.0.1 - 2025-02-03
 
 ### Added